requirements.txt
discord.py>2.0.0
discord-ext-menus==1.1